absorb NotifyUserOfRequestCancellation into PendingRequestForUser
